+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [49.9.4]
 - [SearchBar] Set correct text color.
+- [SearchBar][iOS] Modify background color. 
 - Revert previous commit where things were changed in `IconTintColorHandler`, breaking icon tinting.
 
 ## [49.9.3]